+++ backend/app/models/payroll/tax_bracket.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+from decimal import Decimal
 from typing import Optional
 import uuid
 
@@ -17,9 +18,9 @@ class TaxBracket(Base):
     company_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("companies.id"), nullable=False
     )
-    lower_bound: Mapped[float] = mapped_column(Numeric(15, 2), nullable=False)
-    upper_bound: Mapped[Optional[float]] = mapped_column(Numeric(15, 2))
-    rate_pct: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False)
+    lower_bound: Mapped[Decimal] = mapped_column(Numeric(15, 2), nullable=False)
+    upper_bound: Mapped[Optional[Decimal]] = mapped_column(Numeric(15, 2))
+    rate_pct: Mapped[Decimal] = mapped_column(Numeric(5, 2), nullable=False)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
     created_at = mapped_column(DateTime(timezone=True), server_default=func.now())
